autokicking based on height now uses the current maximum height with or without Wings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7753 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ShotStatistics.cxx
@@ -93,7 +93,7 @@ FlagType* ShotStatistics::getFavoriteFlag() const
 
   // we don't deal with the case where there are two "equally favorite"
   // flags; doesn't really matter
-  for (int i = 0; i < flags.size(); i++) {
+  for (int i = 0; i < (int)flags.size(); i++) {
     if (flags[i].second > greatest.second)
       greatest = flags[i];
   }
@@ -123,7 +123,7 @@ FlagType* ShotStatistics::getBestFlag() const
 
   // we don't deal with the case where there are two "equally best"
   // flags; doesn't really matter
-  for (int i = 0; i < flags.size(); i++) {
+  for (int i = 0; i < (int)flags.size(); i++) {
     if (flags[i].second > greatest.second)
       greatest = flags[i];
   }
